Saiki Kusuo no Sainan (2016) mapping
TVDB lists this series in 2 different ways. One way which was previously used was the 95 specials in season 0 of tvdb which corresponded to the 95 first episodes of the AniDB entry, but this is incomplete and prompted the current change.

The other way, which the mapping has now been changed to are the 24 episodes in season 1 of tvdb where each episode is 5 of the anidb 5 minute episodes.

Since it seems very unlikely the full set of 5 minute episodes will ever be added to tvdb the next best of linking 5 AniDB episodes per TVDB episode appears to be the best solution.

See Issue #251  for the full discussion on this.